Implement an idea from Lex van Roon <r3boot@r3blog.nl.eu.org> to make it possible to delete a list of ident in a given table, rather than iterating ident and droping one ident per table at a time.
This introduce 4 new API function:
- preludedb_delete_alert_from_list() / preludedb_delete_heartbeat_from_list()
- preludedb_delete_alert_from_result_idents() / preludedb_delete_heartbeat_from_result_idents()

Implement workaround to revert to the old implementation in case the format plugin in use does 
not export the necessary function. Using these function to delete list of alert/heartbeat increase 
delete time by about 3000%.

These functions still need to be hooked in preludedb-admin and prewikka so that the user might
see the deletion performance improvement.