Fix warnings and error on ByteA insertion
Using '%s' for inserting data escaped with PQescapeByteaConn() correctly
work with latests PostgreSQL version. However, on version earlier than 9.1,
this generate a warning (HINT:  Use the escape string syntax for backslashes,
e.g., E'\\'. WARNING:  nonstandard use of \\ in a string literal) which
require the user to activate the standard_conforming_strings variable.

On the other end, using the E'%s' escape method is not compatible with
the PQescapeByteaConn() function, we have noticed various error like :
 ERROR:  invalid byte sequence for encoding "UTF8": 0x00
 ERROR:  invalid input syntax for type bytea at character 145

Confirmation of this incompatibility is given by Tom Lane on the PostgreSQL
mailing lists :
http://grokbase.com/t/postgresql/pgsql-general/11b2e9h9xm/postgresql-9-1-1-pqescapebyteaconn-generating-incorrectly-quoted-output

"The result of PQescapeByteaConn has never been meant to be put into an E'' literal.
 You might have managed to get away with it so long as standard_conforming_strings
 was off in the server, but it's on by default in 9.1."

As a result, and since we don't want to force the user to change his PostgreSQL
server configuration, we now use the E'%s' ByteA insertion method in conjonction
with our own escaping function. This function follow the escaping instruction
provided in the PostgreSQL documentation :
http://www.postgresql.org/docs/7.3/static/datatype-binary.html