Error checking fixes, use ENOSYS in place of ENOTSUP
Check _preludedb_plugin_sql_get_row_count() return value with
the prelude_error defined code so that we can workaround the
case where the backend doesn't implement the function.

Additionally, replace any instance of ENOTSUP with ENOSYS. ENOTSUP
is not available everywhere (OpenBSD), so we used to define it to
ENOTSUPP in case it wasn't. This made it harder to check for the
libprelude mapped error code (which could be PRELUDE_ERROR_ENOTSUP
or PRELUDE_ERROR_ENOTSUPP). Using ENOSYS, which seems portable,
make it possible to avoid the problem.

Change-Id: I1e261f2cb2414cf5a2424acf0ac8a731e0c45623