MySQL: fix invalid field length
mysql_fetch_lengths() is only valid for the current row of the result set.
As a result, the lengths pointer we used to hold in preludedb_sql_row_t
wasn't reflecting the given preludedb_sql_row_t fields lengths, but the
current fields lengths.

To fix this problem, we copy the MySQL lengths at preludedb_sql_row_t
creation time. A flexible array member is used to avoid repeated call
to malloc.

Change-Id: Ia06302f5c23e4c99562aa810baab74067d7553ab